(오류) 메뉴 선택시 수량은 올라가지만 total 금액에 추가되지않음

+++ src/main/java/com/cafe/coco/repository/OrderRepository.java
@@ -20,7 +20,7 @@ public interface OrderRepository {
     HashMap<String, Object> modifyMenu(Long pk);
 
     // 금액확인
-    Long total();
+    int total();
